% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{adjust_val}
\alias{adjust_val}
\title{Adjust raw SAGE values based on VIC}
\usage{
adjust_val(val, var_vif, var_vif_threshold)
}
\arguments{
\item{val}{A long vector of unadjusted SAGE values}

\item{var_vif}{A vector of VIF values for each variable. Variables must be in
the same order as in \code{val}.}

\item{var_vif_threshold}{Threshold for \code{var_vif} to consider as "high
colinearity".}
}
\value{
Returns adjusted values, same length as \code{val}.
}
\description{
Adjust raw SAGE values based on VIC
}
