% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{check_df_svic}
\alias{check_df_svic}
\title{Check if input data contains essential information from ShapleyVIC analysis}
\usage{
check_df_svic(df_svic)
}
\arguments{
\item{df_svic}{Input ShapleyVIC output.}
}
\value{
If input passes check, returns it as \code{data.frame}. Otherwise
  returns \code{NULL} and prints data problems.
}
\description{
Check if input data contains essential information from ShapleyVIC analysis
}
