% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{compile_shapley_vic}
\alias{compile_shapley_vic}
\title{Compile ShapleyVIC values and compute overall variable importance}
\usage{
compile_shapley_vic(
  output_dir,
  outcome_type,
  criterion = "loss",
  x = NULL,
  y = NULL,
  x_names_cat = NULL,
  x_names = NULL
)
}
\arguments{
\item{output_dir}{A string indicating path to the output folder created by
the Python library.}

\item{outcome_type}{A string indicating type of outcome. Currently
supports \code{"binary", "continuous", "ordinal"}.}

\item{criterion}{Criterion for defining nearly optimal used in the Python
workflow, which can be "loss" (default, any outcome type), "auc" (binary
outcome only) or "prauc" (binary outcome only).}

\item{x}{A \code{data.frame} of variables in the training set. Required if
\code{x.csv} is not in \code{output_dir}. Ensure \code{x} is identical to 
that used in the Python workflow.}

\item{y}{A vector of the outcome variable in the training set. Required if
\code{y.csv} is not in \code{output_dir}. Ensure \code{y} is identical to 
that used in the Python workflow.}

\item{x_names_cat}{A string vector indicating names of variables in raw data
(i.e., \code{x}) to be treated as categorical. Must be the same as that
used to compute ShapleyVIC values using the Python library.}

\item{x_names}{A string vector indicating alternative variable names to use
in output and plots. If unspecified, variable names in raw data will be
used.}
}
\value{
Returns a \code{ShapleyVIC} object with two elements: (i) a
  \code{data.frame} of ShapleyVIC values from nearly optimal models, adjusted
  based on colinearity among predictors, and (ii) a \code{data.frame} of
  overall variable importance computed from ShapleyVIC values.
}
\description{
Compile ShapleyVIC values and compute overall variable importance
}
