% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shapley_vic.R
\name{compute_loss_bin}
\alias{compute_loss_bin}
\title{Compute logistic loss}
\usage{
compute_loss_bin(coef_vec, x, y)
}
\arguments{
\item{coef_vec}{Regression coefficients of the logistic regression. The first
element should be the intercept term.}

\item{x}{A \code{data.frame} of predictors. Make sure categorical variables
are properly encoded as factors.}

\item{y}{A numeric vector of outcome, with events coded as 1.}
}
\value{
Returns the logistic loss.
}
\description{
Compute logistic loss
}
