% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{compute_meta_interval}
\alias{compute_meta_interval}
\title{Perform meta-analysis to summarise ShapleyVIC results}
\usage{
compute_meta_interval(val, val_sd)
}
\arguments{
\item{val}{A vector of ShapleyVIC values.}

\item{val_sd}{A vector of standard deviation of ShapleyVIC values. Must be
the same length as \code{val}.}
}
\value{
Returns a named vector of summary statistics from the random effects
  model analysis of ShapleyVIC values, including the estimated \code{mean},
  the upper and lower of prediction interval (\code{pred_lower} and
  \code{pred_upper}), the I2 statistic for heterogeneity (\code{hetero_i2})
  and its confidence interval (\code{hetero_i2_lower} and
  \code{hetero_i2_upper}).
}
\description{
Perform meta-analysis to summarise ShapleyVIC results
}
\details{
Fixed effects model is used when all "studies" are from same sample,
  therefore should be estimating exactly the same thing. Random effects model
  thinks different "studies" can have different true value. In traditional
  meta analysis this difference is due to different sample. Here it is due to
  the genuinely different true model reliance in different models.
}
