% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shaply_vic.R
\name{compute_mr_model}
\alias{compute_mr_model}
\title{Compute SAGE model reliance for a model}
\usage{
compute_mr_model(
  model_py,
  model_class = "logit",
  coef_vec = NULL,
  var_names = NULL,
  X,
  y,
  check_convergence = FALSE
)
}
\arguments{
\item{model_py}{A callable model object to pass to Python.}

\item{model_class}{Model class. Currently only supports \code{"logit"} for
logistic regression model.}

\item{coef_vec}{Named numeric vector of model coefficients. The first element
should be the intercept term.}

\item{X}{Numeric matrix of predictors.}

\item{y}{Integer vector of outcome, preferably coded as 1 (event) vs -1 (non-event).}
}
\description{
Compute SAGE model reliance for a model
}
