% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shaply_vic.R
\name{compute_mr_model_reg}
\alias{compute_mr_model_reg}
\title{Compute SAGE model reliance for a model, uing ShapleyReg approximation}
\usage{
compute_mr_model_reg(
  model_class = "logit",
  coef_int,
  coef,
  X_train,
  y_train,
  X,
  y
)
}
\arguments{
\item{model_class}{Model class. Currently only supports \code{"logit"} for
logistic regression model.}

\item{X_train}{Numeric matrix of predictors in training set.}

\item{X}{Numeric matrix of predictors.}

\item{y}{Integer vector of outcome, preferably coded as 1 (event) vs -1 (non-event).}

\item{model_py}{A callable model object to pass to Python.}

\item{coef_vec}{Named numeric vector of model coefficients. The first element
should be the intercept term.}
}
\description{
Compute SAGE model reliance for a model, uing ShapleyReg approximation
}
