% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shapley_vic.R
\name{compute_shapley_vic}
\alias{compute_shapley_vic}
\title{Compute ShapleyVIC values for nearly optimal models}
\usage{
compute_shapley_vic(
  model_py,
  model_colinear = NULL,
  coef_mat,
  perf_metric,
  x_test,
  y_test,
  var_names = NULL,
  n_cores,
  output_folder = NULL
)
}
\arguments{
\item{model_py}{A Python callable model object that has a
\code{predict_proba} function, or the path to a file that stores the model.}

\item{model_colinear}{A model train from training set to detect colinearity
of predictors (e.g., a logistic regression). This is relevant to any model
that assumes a linear function for predictors.}

\item{coef_mat}{A numeric matrix or \code{data.frame} of coefficients of
nearly optimal models, where each row corresponds to a model and each
column corresponds to a variable (or a category of a categorical variable).}

\item{perf_metric}{Performance metric of each model in \code{coef_mat}.}

\item{x_test}{A \code{data.frame} of predictors from the test set. Make sure
categorical variables are properly encoded as factors.}

\item{var_names}{String vector of variable names (not the names of regression
coefficients, if categorical variables are involved). If unspecified,
column names of \code{x_test} will be used.}

\item{n_cores}{Number of cores to use for parallel computing.}
}
\description{
Compute ShapleyVIC values for nearly optimal models
}
