% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_compas_models}
\alias{df_compas_models}
\title{350 nearly optimal logistic regression models for predicting COMPAS data}
\format{
An object of class \code{data.frame} with 350 rows and 8 columns.
}
\usage{
df_compas_models
}
\description{
A \code{data.frame} with 350 rows (each corresponding to a
  model) and 8 columns, where the first 7 columns correspond to the
  regression coefficients and the 8th column correspond to the ratio of
  logistic loss to the minimum loss (i.e., logistic loss of model trained
  from training set).Code used to generated these 350 models is included in
  Example section.
}
\examples{
data("df_compas")
set.seed(1234)
i_test <- sample(1:nrow(df_compas), size = round(nrow(df_compas) * 0.1),
                 replace = FALSE)
# i_test was later recorded as the `train_test` column
df_train <- df_compas[-i_test, ]
df_test <- df_compas[i_test, ]
m_optim_r <- glm(y ~ ., data = df_train, family = "binomial")
df_compas_models <- ShapleyVIC::draw_models(
  coef_optim = coef(m_optim_r),
  coef_optim_var = vcov(m_optim_r),
  x = df_train[, -1], y = df_train$y,
  M = 800, u1 = 0.5, u2 = 80, epsilon = 0.05,
  n_final = 350
)
head(df_compas_models)
}
\seealso{
\code{\link{df_compas}}, \code{\link{df_compas_shapley_vic}}
}
\keyword{datasets}
