% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_models.R
\name{draw_models}
\alias{draw_models}
\title{Generate models from Rashomon set}
\usage{
draw_models(
  coef_optim,
  coef_optim_var,
  design_mat,
  y,
  M = 800,
  u1 = 0.5,
  u2 = 20,
  epsilon = 0.05,
  n_final = NULL
)
}
\description{
Generate models from Rashomon set
}
