% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{draw_violins}
\alias{draw_violins}
\title{Draw colored violin plot for model reliance.}
\usage{
draw_violins(
  var_names,
  val,
  perf_metric,
  smaller_is_better = TRUE,
  title = "",
  colored = TRUE,
  my_width = 0.35,
  perf_metric_breaks = c(1.01, 1.02, 1.03, 1.04)
)
}
\arguments{
\item{var_names}{Factor of variable names. Must be a factor.}

\item{val}{Model reliance.}

\item{title}{Title of violin plot. Default is an empty string.}

\item{colored}{Whether the violins should be colored by model loss.}

\item{loss_ratio}{Ratio of loss to optimal loss, to determine color.}
}
\description{
Draw colored violin plot for model reliance.
}
