% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_setup.R
\name{install_matplotlib}
\alias{install_matplotlib}
\title{R wrapper for installing matplotlib python library}
\usage{
install_matplotlib(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows. See \code{\link[reticulate]{py_install}}.}

\item{conda}{The path to a conda executable. Use "auto" to allow reticulate
to automatically find an appropriate conda binary. See
\code{\link[reticulate]{py_install}}.}
}
\description{
R wrapper for installing matplotlib python library
}
