% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_setup.R
\name{install_python_lib}
\alias{install_python_lib}
\title{Install required Python libraries}
\usage{
install_python_lib(install_shap = TRUE)
}
\arguments{
\item{install_shap}{Whether the SHAP Python library should be installed.
Default is \code{TRUE}. See Details.}
}
\description{
Installs Python libraries sage-importance, sklearn, pandas and
  shap (if \code{install_shap = TRUE}) to current Python inspector specified
  through in RStudio menu.
}
\details{
Currently SHAP cannot be installed for Mac Silicon machines due to
  compatibility issues. To install the SHAP Python library on Windows
  machines, users must first install the Microsoft Visual C++ (version 14.0
  or later), which is not available by default. If users prefer not to or are
  unable to install this tool, they can choose not to install the SHAP Python
  library and not to use the \code{compute_shap_value} function.
}
