% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shaply_vic.R
\name{logit_model_python}
\alias{logit_model_python}
\title{Fit logistic regression model using sklearn python library}
\usage{
logit_model_python(x_train, y_train, solver = "liblinear")
}
\arguments{
\item{x_train}{Numeric design matrix of predictors.}

\item{y_train}{Integer vector of outcome, preferably coded as 1 (event) vs -1
(non-event).}

\item{solver}{Solver to use by the python function.}
}
\value{
Returns a callable python model object to be used by
  \code{\link{compute_sage_value}}.
}
\description{
Fit logistic regression model using sklearn python library
}
