% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{make_violin_legend}
\alias{make_violin_legend}
\title{Create figure legend for violin plot}
\usage{
make_violin_legend(perf_metric_breaks, col_vec)
}
\arguments{
\item{perf_metric_breaks}{Thresholds for performance metrics.}

\item{col_vec}{Vector of color to use for each performance interval.}
}
\description{
Create figure legend for violin plot
}
