% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_models.R
\name{mark_eligibility}
\alias{mark_eligibility}
\title{Mark whether newly generated models are eligible}
\usage{
mark_eligibility(coef_df, coef_optim, x, y, criterion = "loss", epsilon = 0.05)
}
\arguments{
\item{coef_df}{A \code{data.frame} of models generated from
\code{\link{draw_models_initial}}.}

\item{coef_optim}{Regression coefficients of the optimal logistic regression
(i.e., the model that minimized logistic loss).}

\item{x}{A \code{data.frame} of predictors from the training set. Make
sure categorical variables are properly encoded as factors.}

\item{y}{A numeric vector of outcome, with events coded as 1.}

\item{criterion}{Criterion for determining model eligibility. Currently
supporting \code{"loss"}.}

\item{epsilon}{Nearly optimal models are defined as models with logistic loss
less than (1+epsilon) times the minimum loss.}
}
\value{
Returns \code{coef_df} with 2 additional columns, \code{perf_metric}
  and \code{eligible}.
}
\description{
Mark whether newly generated models are eligible
}
