% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shapley_vic.R
\name{model_matrix_no_intercept}
\alias{model_matrix_no_intercept}
\title{Create a design matrix without the intercept term}
\usage{
model_matrix_no_intercept(formula, data)
}
\arguments{
\item{formula}{Formula.}

\item{data}{A \code{data.frame} of predictors. Make sure categorical
variables are properly encoded as factors.}
}
\value{
Returns a numeric matrix.
}
\description{
Create a design matrix without the intercept term
}
