% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{plot.ShapleyVIC}
\alias{plot.ShapleyVIC}
\title{Plot ShapleyVIC findings}
\usage{
\method{plot}{ShapleyVIC}(x, ...)
}
\arguments{
\item{x}{ShapleyVIC object generated using \code{\link{compile_shapley_vic}}.}

\item{...}{Not implemented}
}
\value{
Displays bar and violin plots and returns the two plot objects as a
  list.
}
\description{
Plot ShapleyVIC findings
}
