% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{plot_bars}
\alias{plot_bars}
\title{Make bar plots for ShapleyVIC findings}
\usage{
plot_bars(x, title = NULL, subtitle = NULL)
}
\arguments{
\item{x}{ShapleyVIC object generated using \code{\link{compile_shapley_vic}}.}

\item{title}{Title of the bar plot (optional).}

\item{subtitle}{Subtitle of the bar plot (optional).}
}
\value{
Returns a bar plot that can be further edited using ggplot2.
}
\description{
Make bar plots for ShapleyVIC findings
}
