% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{plot_violin}
\alias{plot_violin}
\title{Make violin plot for ShapleyVIC findings}
\usage{
plot_violin(x, title = NULL, plot_theme = NULL)
}
\arguments{
\item{x}{ShapleyVIC object generated using \code{\link{compile_shapley_vic}}.}

\item{title}{Title of violin plot (optional).}

\item{plot_theme}{\code{ggplot theme} to apply to the violin plot, if any.}
}
\value{
Returns a violin plot. Use \code{plot_theme} to specify
  \code{theme()} settings.
}
\description{
Make violin plot for ShapleyVIC findings
}
