% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{print.ShapleyVIC}
\alias{print.ShapleyVIC}
\title{Print ShapleyVIC object}
\usage{
\method{print}{ShapleyVIC}(x, digits = 5, ...)
}
\arguments{
\item{x}{ShapleyVIC object generated using \code{\link{compile_shapley_vic}}.}

\item{digits}{Number of digits to print for numeric values. Default is 5.}

\item{...}{Not implemented}
}
\description{
Print ShapleyVIC object
}
