% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{rank_vars}
\alias{rank_vars}
\title{Rank variables based on pairwise comparison of model reliance for one model}
\usage{
rank_vars(val, val_sd = NULL, ties.method = "min")
}
\arguments{
\item{val}{A numeric vector of model reliance.}

\item{val_sd}{A numeric vector of standard deviations of model reliance (with
the same length as \code{val}), if available.}

\item{ties.method}{How to handle tied ranks. Default is \code{"min"}. See
\code{\link{rank}}.}
}
\value{
Returns an integer vector of ranks of variables in descending order.
}
\description{
Rank variables based on pairwise comparison of model reliance for one model
}
