% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shapley_vic.R
\name{replace_coef}
\alias{replace_coef}
\title{Replace estimated coefficients in a Python object with prespeficied values.}
\usage{
replace_coef(model_py, coef_vec)
}
\description{
Replace estimated coefficients in a Python object with prespeficied values.
}
