% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{set_color_template}
\alias{set_color_template}
\title{Set color template for bar plots}
\usage{
set_color_template(sign_vec, color_template = NULL)
}
\arguments{
\item{sign_vec}{A vector of -1, 0, 1.}

\item{color_template}{A vector of 3 colours corresponding to -1, 0 and 1.}
}
\description{
Set color template for bar plots
}
