% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{summ_shapley_vic}
\alias{summ_shapley_vic}
\title{Summarize ShapleyVIC values}
\usage{
summ_shapley_vic(val, val_sd, var_names)
}
\arguments{
\item{val}{A vector of ShapleyVIC values from all models evaluated.}

\item{val_sd}{A vector of standard deviation of ShapleyVIC values from all
models evaluated. Must be the same length as \code{val}. If not available,
assign to \code{NULL}.}

\item{var_names}{A vector of variable names.}
}
\value{
Returns a \code{data.frame} of variable names, their average
  ShapleyVIC values, and the lower and upper bounds of 95\% prediction
  intervals (\code{NA} if \code{val_sd = NULL}).
}
\description{
Summarize ShapleyVIC values
}
