% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{summarise_shapley_vic}
\alias{summarise_shapley_vic}
\title{Summarize ShapleyVIC values}
\usage{
summarise_shapley_vic(
  val,
  val_sd,
  var_names,
  var_vif = NULL,
  var_vif_threshold = 2
)
}
\arguments{
\item{val}{A vector of ShapleyVIC values from all models evaluated.}

\item{val_sd}{A vector of standard deviation of ShapleyVIC values from all
models evaluated. Must be the same length as \code{val}. If not available,
assign to \code{NULL}.}

\item{var_names}{A vector of variable names.}

\item{var_vif}{If \code{val} is the unadjusted SAGE values, specify a vector
of variance inflation factors (VIF) for each variable to obtain adjusted
ShapleyVIC values. Otherwise leave as \code{NULL}.}

\item{var_vif_threshold}{Threshold for adjusting SAGE values based on VIF.
Default is 2.}
}
\value{
Returns a \code{data.frame} of variable names, their average
  ShapleyVIC values, and the lower and upper bounds of 95\% prediction
  intervals (\code{NA} if \code{val_sd = NULL}).
}
\description{
Summarize ShapleyVIC values
}
