% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{vars_to_exclude}
\alias{vars_to_exclude}
\title{Identify variables with zero importance in all models, to exclude from plots}
\usage{
vars_to_exclude(df_bar)
}
\arguments{
\item{df_bar}{Data frame of overall variable importance.}
}
\value{
Returns a vectors of variables to exclude (vector of zero length if
  none to exclude).
}
\description{
Identify variables with zero importance in all models, to exclude from plots
}
