% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{compute_meta_interval}
\alias{compute_meta_interval}
\title{Perform random effects meta-analysis to summarize ShapleyVIC results}
\usage{
compute_meta_interval(val, val_sd)
}
\arguments{
\item{val}{A vector of ShapleyVIC values.}

\item{val_sd}{A vector of standard deviation of ShapleyVIC values. Must be
the same length as \code{val}.}
}
\value{
Returns a named vector of summary statistics from the random effects
  model analysis of ShapleyVIC values, including the estimated \code{mean},
  the upper and lower of prediction interval (\code{pred_lower} and
  \code{pred_upper}).
}
\description{
Perform random effects meta-analysis to summarize ShapleyVIC results
}
