% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shapley_vic.R
\name{compute_shapley_vic}
\alias{compute_shapley_vic}
\title{Compute ShapleyVIC values for nearly optimal models}
\usage{
compute_shapley_vic(
  model_py,
  coef_mat,
  perf_metric,
  var_vif = NULL,
  var_vif_threshold = 2,
  x_test,
  y_test,
  var_names = NULL,
  output_folder = NULL,
  n_cores
)
}
\arguments{
\item{model_py}{A Python callable model object that has a
\code{predict_proba} function, or the path to a file that stores the model.}

\item{coef_mat}{A numeric matrix or \code{data.frame} of coefficients of
nearly optimal models, where each row corresponds to a model and each
column corresponds to a variable (or a category of a categorical variable).}

\item{perf_metric}{Performance metric of each model in \code{coef_mat}.}

\item{var_vif}{Variable inflation factor (VIF) for each variable (optional).
If provided, absolute SAGE values will be used instead of unadjusted SAGE
values as final ShapleyVIC values for variables with
\code{var_vif>var_vif_threshold}. If not specified, unadjusted SAGE values
will be saved as ShapleyVIC values, which can be inspected and adjusted in
the summarisation and visualisation steps later.}

\item{var_vif_threshold}{Threshold for adjusting SAGE values based on VIF.
Default is 2.}

\item{x_test}{A \code{data.frame} of predictors from the test set. Make sure
categorical variables are properly encoded as factors.}

\item{y_test}{A numeric vector of outcome from the test set, with events
coded as 1.}

\item{var_names}{String vector of variable names (not the names of regression
coefficients, if categorical variables are involved). If unspecified,
column names of \code{x_test} will be used.}

\item{output_folder}{Folder to save ShapleyVIC values from individual models
(optional).}

\item{n_cores}{Number of cores to use for parallel computing.}
}
\description{
Compute ShapleyVIC values for nearly optimal models
}
