% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_compas_shapley_vic}
\alias{df_compas_shapley_vic}
\title{ShapleyVIC values from the 350 nearly optimal models for COMPAS data}
\format{

}
\usage{
df_compas_shapley_vic
}
\description{
ShapleyVIC values from the 350 nearly optimal models for COMPAS data
}
\seealso{
\code{\link{df_compas}}, \code{\link{df_compas_models}}
}
\keyword{datasets}
