% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{draw_bars}
\alias{draw_bars}
\title{Draw bar plot for model reliance.}
\usage{
draw_bars(
  val,
  val_sd = NULL,
  val_lower = NULL,
  val_upper = NULL,
  var_names = NULL,
  labels = NULL,
  title = "",
  subtitle = "",
  sign_vec = NULL,
  color_template = NULL
)
}
\description{
Draw bar plot for model reliance.
}
