% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shapley_vic.R
\name{logit_model_python}
\alias{logit_model_python}
\title{Fit logistic regression model using sklearn python library}
\usage{
logit_model_python(x_train, y_train)
}
\arguments{
\item{x_train}{A \code{data.frame} of predictors from the training set. Make
sure categorical variables are properly encoded as factors.}

\item{y_train}{A numeric vector of outcome, with events coded as 1.}
}
\value{
Returns a callable python model object to be used by
  \code{\link{compute_shap_value}} and \code{\link{compute_sage_value}},
  where the estimated coefficients will be identical to those from the
  \code{glm} function.
}
\description{
Fit logistic regression model using sklearn python library
}
