% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{prep_data_zebra}
\alias{prep_data_zebra}
\title{Prepare data for making colored violin plot}
\usage{
prep_data_zebra(
  var_names,
  val,
  perf_metric,
  perf_metric_breaks = c(1.01, 1.02, 1.03, 1.04),
  my_width = 0.35
)
}
\arguments{
\item{var_names}{Factor of variable names. Must be a factor.}

\item{val}{Model reliance.}

\item{loss_ratio}{Ratio of loss to optimal loss, to determine color.}
}
\description{
Prepare data for making colored violin plot
}
