% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{rank_variables}
\alias{rank_variables}
\title{Rank variables based on pairwise comparison of model reliance}
\usage{
rank_variables(x, summarise = FALSE, filter = TRUE, as_vector = FALSE)
}
\arguments{
\item{x}{ShapleyVIC object generated using \code{\link{compile_shapley_vic}}.}

\item{summarise}{If \code{TRUE}, ranks will be summarised across models by
taking the numeric average. Default is \code{FALSE}, where ranks for
individual models will be returned.}

\item{filter}{Whether to filter out variables with non-significant overall
importance. Default is \code{TRUE}.}

\item{as_vector}{If \code{summarise = TRUE}, whether to return variable ranks
as a named vector (if \code{TRUE}, and using raw variable names), or as a
\code{data.frame} (if \code{FALSE}, default).}
}
\value{
Returns an integer vector of ranks of variables for each model (in
  descending order). Model reliance (possibly with 95\% confidence interval)
  is compared between all possible pairs of variables. Variables with model
  reliance significantly than the other variable in more pairwise comparisons
  are ranked higher.
}
\description{
Rank variables based on pairwise comparison of model reliance
}
