% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_visualise.R
\name{summarise_shapley_vic}
\alias{summarise_shapley_vic}
\title{Summarize ShapleyVIC values}
\usage{
summarise_shapley_vic(val, val_sd, var_names)
}
\arguments{
\item{val}{A vector of ShapleyVIC values from all models evaluated.}

\item{val_sd}{A vector of standard deviation of ShapleyVIC values from all
models evaluated. Must be the same length as \code{val}. If not available,
assign to \code{NULL}.}

\item{var_names}{A vector of variable names corresponding to each ShapleyVIC
value. Should be the same length as \code{val}. Will be recycled if shorter
than \code{val}.}
}
\value{
Returns a \code{data.frame}.
}
\description{
Summarize ShapleyVIC values
}
